optimized directory index creation. Fixed bug using nFile inst iThisFile
Was incorrectly using the iThisFileInDir logical index array on the FDList cell array. This caused the routine to strip out good rows in the FDList on each execution. Renamed the iterator `i` as `nFile` to help differentiate that is refers only to the `files` `fileNames` and `fileDates` arrays, as derived from the `dir()` command.