Updated import parser: no more line count, encoding detection on macos
also increased the chunk size. Maybe that should be a parameter, but for now this is good. Learned that matfile() can't write data to saved variables in struct fields, so the current FD data structure is unsupported. That would have made things RADICALLY better.

Perhaps a new FD data type can me created with a version tag, so the old data are still accessible. The new one can simply keep the timeseries data in vectors, for easy streaming to disk.